% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyEmptyImageDimensions.R
\name{applyEmptyImageDimensions}
\alias{applyEmptyImageDimensions}
\alias{apply_empty_dim}
\title{Apply Subsetting from Empty Image Dimensions}
\usage{
applyEmptyImageDimensions(img, inds, reorient = FALSE)

apply_empty_dim(img, inds, reorient = FALSE)
}
\arguments{
\item{img}{nifti object}

\item{inds}{indices of subset from \code{\link{getEmptyImageDimensions}} or
\code{\link{dropEmptyImageDimensions}}.}

\item{reorient}{Should image be reoriented if a filename?}
}
\value{
Object of class \code{nifti}
}
\description{
Simple wrapper for subsetting an image with indices, 
dropping empty dimensions.
}
\note{
\code{apply_empty_dim} is a shorthand for 
\code{applyEmptyImageDimensions} with all the same arguments.
}
\seealso{
\code{\link{getEmptyImageDimensions}}, 
\code{\link{dropEmptyImageDimensions}}
}

