% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applytopup.R
\name{applytopup}
\alias{applytopup}
\alias{apply_topup}
\alias{fsl_applytopup}
\title{applytopup - calling FSL applytopup}
\usage{
applytopup(infile, datain, index, topup_files, out = NULL,
  method = c("lsr", "jac"), interp = c("spline", "trilinear"),
  verbose = TRUE)

apply_topup(...)

fsl_applytopup(...)
}
\arguments{
\item{infile}{list of names of input image (to be corrected)}

\item{datain}{name of text file with PE directions/times}

\item{index}{list of indicies into --datain of 
the input image (to be corrected)}

\item{topup_files}{name of field/movements (from topup)}

\item{out}{basename for output (warped) image}

\item{method}{Use jacobian modulation (jac) or 
least-squares resampling (lsr), default=lsr.}

\item{interp}{Image interpolation model, trilinear or spline. Default spline}

\item{verbose}{Print diagnostic information while running}

\item{...}{arguments passed to \code{topup} if using \code{fsl_topup}}
}
\description{
A tool for applying and correcting 
estimated susceptibility induced distortions
}
