% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_file.R
\docType{methods}
\name{aux.file-methods}
\alias{aux.file-methods}
\alias{aux.file,character-method}
\title{Extract Image aux.file attribute}
\usage{
\S4method{aux.file}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
aux_file method for character types
}
