% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaker.R
\name{breaker}
\alias{breaker}
\title{Find Breaks for nifti Image Ploting}
\usage{
breaker(x, zlim, col, breaks = NULL)
}
\arguments{
\item{x}{Object of class \code{nifti}}

\item{zlim}{A user-specified \code{zlim}.  If \code{NULL}, will calculate how 
\code{\link{ortho2}} would calculate \code{zlim}}

\item{col}{colors to be plotted.  Only used for \code{length(col)}, so can be a
vector of length cols to be plotted}

\item{breaks}{if \code{!is.null(breaks)}, then will calculate breaks.  Otherwise will
return this breaks vector}
}
\value{
Vector of length 2

If \code{breaks = NULL}, then vector of \code{length(col) + 1}, 
otherwise returns \code{breaks}
}
\description{
Helper function for plotting - returns breaks for \code{\link[graphics]{image}} 
plot function for object of class \code{nifti}
}

