% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cal_max.R
\docType{methods}
\name{cal.min,character-method}
\alias{cal.min,character,character-method}
\alias{cal.min,character-method}
\title{Extract or Replace NIfTI/Analyze Min or Max Values}
\usage{
\S4method{cal.min}{character}(object)
}
\description{
Extract or Replace NIfTI/Analyze Min or Max Values
}

