% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nii.stub.R
\name{cal_img}
\alias{cal_img}
\title{Set Max/Min for nifti object by range of data}
\usage{
cal_img(img, infok = TRUE)
}
\arguments{
\item{img}{nifti object}

\item{infok}{Indicator if \code{Inf} and \code{-Inf} are OK.  Default is
\code{TRUE}.  If \code{FALSE} and max or min is infinity, then
\code{cal_min}
or \code{cal_max} is set to infinity (neg or pos), respectively.}
}
\value{
object of type nifti
}
\description{
Rescales image cal_max and cal_min slots to be the
max and min,
respectively, of an object of class nifti, with \code{na.rm=TRUE}.
This is so that
when images are rendered/written, the values correspond to those
in the array (stored in .Data slot) are plotted on correct grayscale
and no error is
given by \code{writeNIfTI}.
}

