% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mask.R
\name{check_mask_fail}
\alias{check_mask_fail}
\title{Check Mask is Binary, Fail otherwise}
\usage{
check_mask_fail(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{check_mask}}}
}
\value{
Either will error if conditions not met or an invisible \code{NULL}
}
\description{
Determine if only values in a mask are 0/1.  Will error otherwise.
}
\examples{
arr = array(rbinom(1000, size = 1, prob = 0.2), dim = c(10,10,10))
nim = oro.nifti::nifti(arr)
check_mask_fail(nim)
}

