% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outfile.R
\name{check_outfile}
\alias{check_outfile}
\title{Check output filename}
\usage{
check_outfile(outfile, retimg, fileext = ".nii.gz")
}
\arguments{
\item{outfile}{output filename or NULL}

\item{retimg}{Should an image be returned}

\item{fileext}{a non-empty character vector giving the file extension}
}
\value{
Filename of output file or a temporary filename
}
\description{
This function checks if an output filename is not NULL 
in conjunction
whether the user would like to return an image
}

