% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkimg.R
\docType{methods}
\name{checkimg-methods}
\alias{checkimg}
\alias{checkimg,character-method}
\alias{checkimg,list-method}
\alias{checkimg,nifti-method}
\alias{checkimg-methods}
\title{Force object to filename}
\usage{
checkimg(file, ...)

\S4method{checkimg}{nifti}(file, ...)

\S4method{checkimg}{character}(file, ...)

\S4method{checkimg}{list}(file, ...)
}
\arguments{
\item{file}{character or \code{nifti} object}

\item{...}{options passed to \code{\link{tempimg}}}
}
\value{
character filename of image or temporary nii, 
with .nii extension
}
\description{
Ensures the output to be a character filename (or vector) from an input
image or \code{nifti}.
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}

