% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslhd.R
\name{checkimg}
\alias{checkimg}
\title{Check if filename is character or nifti object}
\usage{
checkimg(file, ...)
}
\arguments{
\item{file}{character or nifti object}

\item{...}{options passed to \code{\link{tempimg}}}
}
\value{
character filename of image or temporary nii
}
\description{
Check if filename is character or nifti object
}

