% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checknii.R
\docType{methods}
\name{checknii-methods}
\alias{checknii}
\alias{checknii,character-method}
\alias{checknii,list-method}
\alias{checknii,nifti-method}
\alias{checknii-methods}
\title{Force object to filename with .nii extension}
\usage{
checknii(file, ...)

\S4method{checknii}{nifti}(file, ...)

\S4method{checknii}{character}(file, ...)

\S4method{checknii}{list}(file, ...)
}
\arguments{
\item{file}{character or \code{nifti} object}

\item{...}{options passed to \code{\link{checkimg}}}
}
\value{
character filename of image or temporary nii,
with .nii extension
}
\description{
Force object to filename with .nii extension
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}

