% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checknii.R
\name{checknii}
\alias{checknii}
\title{Force object to filename with .nii extension}
\usage{
checknii(file, ...)
}
\arguments{
\item{file}{character or \code{nifti} object}

\item{...}{options passed to \code{\link{tempimg}}}
}
\value{
character filename of image or temporary nii,
with .nii extension
}
\description{
Force object to filename with .nii extension
}

