% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ortho2.R
\name{colorbar}
\alias{colorbar}
\title{Add a colorbar to an ortho2 object}
\usage{
colorbar(breaks, col, text.col = "white")
}
\arguments{
\item{breaks}{a set of finite numeric breakpoints for the colours(see \code{\link{image}}}

\item{col}{a list of colors (see \code{\link{image}}}

\item{text.col}{axis and text label color}
}
\value{
A plot
}
\description{
Adds a series of colors mapped to a value
}
\note{
Much of this was taken from \code{vertical.image.legend} from
the \code{aqfig} package
}

