% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\docType{methods}
\name{data_type-methods}
\alias{data_type,character-method}
\alias{data_type-methods}
\title{Extract Image data_type attribute}
\usage{
\S4method{data_type}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
data_type method for character types
}

