% Generated by roxygen2 (4.0.1): do not edit by hand
\name{datatype}
\alias{datatype}
\title{Change Data type for img}
\usage{
datatype(img, datatype = NULL, bitpix = NULL, trybyte = TRUE)
}
\arguments{
\item{img}{nifti object (or character of filename)}

\item{datatype}{(NULL) character of datatype see
\code{\link{convert.datatype}}}

\item{bitpix}{(NULL) character of bitpix see
\code{\link{convert.bitpix}}}

\item{trybyte}{(logical) Should you try to make a byte (UINT8) if image in
c(0, 1)?}
}
\value{
object of type nifti
}
\description{
Tries to figure out the correct datatype for image.  Useful
for image masks - makes them binary if
}

