% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cal_max.R
\docType{methods}
\name{descrip,character-method}
\alias{descrip,character,character-method}
\alias{descrip,character-method}
\title{Extract or Replace NIfTI/Analyze Description}
\usage{
\S4method{descrip}{character}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti}}
}
\description{
Extract or Replace NIfTI/Analyze Description
}

