% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dim_.R
\docType{methods}
\name{dim_-methods}
\alias{dim_-methods}
\title{Extract Image dim_ attribute}
\usage{
\S4method{dim_}{character}(object)
}
\arguments{
\item{object}{is an object of class \code{character}}
}
\description{
Methods that extract the ``dim_'' slot from the
NIfTI/ANALYZE header.
}

