% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fsl.R
\name{download_fsl}
\alias{download_fsl}
\title{Download FSL}
\usage{
download_fsl(os = c("macosx", "redhat5", "redhat6", "centos5", "centos6",
  "debian", "ubuntu"), outdir = tempdir(), overwrite = TRUE, ...)
}
\arguments{
\item{os}{Operating system}

\item{outdir}{Output directory for tarball}

\item{overwrite}{If \code{file.path(outdir, tarball_name)} exists,
should it be overwritten?}

\item{...}{Arguments to pass to \code{\link{download.file}}}
}
\value{
Filename of destination file
}
\description{
Download FSL Tarball
}

