% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drop_img_dim}
\alias{drop_img_dim}
\title{Drop Image Dimension}
\usage{
drop_img_dim(img)
}
\arguments{
\item{img}{nifti object}
}
\value{
Object of class nifti
}
\description{
Drops a dimension of an image that has one-dimension and
sets respective values to 0 in pixdim or 1 in dim
}

