% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy_correct.R
\name{eddy_correct}
\alias{eddy_correct}
\title{Eddy Current Correction}
\usage{
eddy_correct(infile, outfile = NULL, retimg = TRUE, reference_no = 0, ...)
}
\arguments{
\item{infile}{input filename of 4D image.}

\item{outfile}{Output filename}

\item{retimg}{(logical) return image of class nifti}

\item{reference_no}{Set the volume number for the reference volume 
that will be used as a target to register all other volumes to. 
(default=0, i.e. the first volume)}

\item{...}{Additional arguments passed to \code{\link{fslcmd}}}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{eddy_correct} from FSL for 
DTI Processing
}
