% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_form.R
\name{enforce_form}
\alias{enforce_form}
\title{Enforce Either Qform or Sform is set}
\usage{
enforce_form(file, ...)
}
\arguments{
\item{file}{(character) image filename or character 
of class nifti}

\item{...}{additional arguments to pass to 
\code{\link{getForms}}}
}
\value{
A character filename
}
\description{
Enforce Either Qform or Sform is set
}
\examples{
if (have_fsl()) {
res = enforce_form(mni_fname())
}
}
