% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fast.R
\name{fast.help}
\alias{fast.help}
\title{FAST help}
\usage{
fast.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fast}'s help
}
\examples{
if (have.fsl()){
 fast.help()
}
}

