% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_imgext.R
\name{file_imgext}
\alias{file_imgext}
\title{Get Image file extension}
\usage{
file_imgext(file, withdot = TRUE)
}
\arguments{
\item{file}{Vector of character filenames}

\item{withdot}{Should the extension begin with \code{"."}?}
}
\value{
Vector of extensions.  If \code{withdot = FALSE}, then will
return \code{nii} instead of \code{.nii}
}
\description{
Get the image file extension, either .nii, .hdr, .nii.gz, 
or .hdr.gz
}

