% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_img.R
\docType{methods}
\name{finite_img-methods}
\alias{finite_img}
\alias{finite_img,character-method}
\alias{finite_img,finite_img-method}
\alias{finite_img,list-method}
\alias{finite_img,nifti-method}
\alias{finite_img-methods}
\title{Finite Image}
\usage{
finite_img(img, replace = 0)

\S4method{finite_img}{nifti}(img, replace = 0)

\S4method{finite_img}{character}(img, replace = 0)

\S4method{finite_img}{list}(img, replace = 0)
}
\arguments{
\item{img}{character path of image or 
an object of class \code{nifti}, or list of images}

\item{replace}{Value to replace non-finite values to}
}
\value{
nifti object
}
\description{
Simple wrapper for setting non-finite values to zero
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}

