% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_img.R
\name{finite_img}
\alias{finite_img}
\title{Replace non-finite values}
\usage{
finite_img(x, replace = 0)
}
\arguments{
\item{x}{is either a character name for the image or an object of class nifti}

\item{replace}{Value to replace non-finite values}
}
\value{
object of class nifti
}
\description{
Replaces non-finite values of nifti image with a replacement
}

