% Generated by roxygen2 (4.0.2): do not edit by hand
\name{flirt.help}
\alias{flirt.help}
\alias{flirt_apply.help}
\title{FLIRT help}
\usage{
flirt.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{flirt}'s help
}
\examples{
if (have.fsl()){
 flirt.help()
}
}

