% Generated by roxygen2 (4.0.2): do not edit by hand
\name{flirt_apply}
\alias{flirt_apply}
\title{Apply Warp from FLIRT}
\usage{
flirt_apply(infile, reffile, initmat, outfile = NULL, retimg = FALSE,
  reorient = FALSE, intern = TRUE, opts = "", verbose = TRUE, ...)
}
\arguments{
\item{infile}{(character) input filename}

\item{reffile}{(character) reference image to be registered to}

\item{initmat}{(character) Matrix of transformation}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{opts}{(character) additional options to FLIRT}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
character or logical depending on intern
}
\description{
This function applies a matrix from \code{\link{flirt}}
to other images
}

