% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fnirt_with_affine_apply}
\alias{fnirt_with_affine_apply}
\title{<brief desc>}
\usage{
fnirt_with_affine_apply(infile, reffile, flirt.omat = NULL,
  flirt.outfile = NULL, fnirt.warpfile = NULL, outfile = NULL,
  retimg = FALSE, reorient = FALSE, intern = TRUE, flirt.opts = "",
  opts = "", verbose = TRUE, ...)
}
\arguments{
\item{infile}{(character) input filename}

\item{reffile}{(character) reference image to be registered to}

\item{flirt.omat}{(character) Filename of output affine matrix}

\item{flirt.outfile}{(character) Filename of output affine-registered image}

\item{fnirt.warpfile}{(character) Filename of warp image from
\code{\link{fnirt}}}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{flirt.opts}{(character) additional options to FLIRT}

\item{opts}{(character) additional options to FNIRT}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
character or logical depending on intern
}
\description{
<full description>
}
\seealso{
fnirt_with_affine
}

