% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fsl_anat.help}
\alias{fsl_anat.help}
\title{fsl_anat help}
\usage{
fsl_anat.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fsl_anat}'s help
}
\examples{
if (have.fsl()){
 fsl_anat.help()
}
}

