% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applywarp.R
\name{fsl_applywarp}
\alias{fsl_applywarp}
\title{Apply Warp from FNIRT}
\usage{
fsl_applywarp(infile, reffile, warpfile, outfile = NULL, retimg = TRUE,
  reorient = FALSE, intern = FALSE, opts = "", verbose = TRUE, ...)
}
\arguments{
\item{infile}{(character) input filename}

\item{reffile}{(character) reference image to be registered to}

\item{warpfile}{(character) reference image to be registered to}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{opts}{(character) additional options to FLIRT}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
character or logical depending on intern
}
\description{
This function applies a coefficient map from \code{\link{fnirt}}
to other images
}

