% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avscale.R
\name{fsl_avscale}
\alias{fsl_avscale}
\alias{avscale}
\title{Scale Affine Matrix using \code{avscale}}
\usage{
fsl_avscale(file, volume = NULL, parsed = TRUE, verbose = TRUE)

avscale(...)
}
\arguments{
\item{file}{(character) matrix filename}

\item{volume}{(character) non-reference volume filename or nifti image}

\item{parsed}{(logical) should \code{\link{parse_avscale}} be run 
after?}

\item{verbose}{(logical) print out command before running}

\item{...}{not used, but used for duplicating \code{avscale} as alias}
}
\value{
Character of information from avscale
}
\description{
This function calls \code{avscale} to get individual
matrices for FSL
}
