% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fsl_biascorrect}
\alias{fsl_biascorrect}
\title{FSL Bias Correct}
\usage{
fsl_biascorrect(file, outfile = NULL, retimg = FALSE, opts = "",
  remove.seg = TRUE, ...)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{opts}{(character) operations to be passed to \code{fast}}

\item{remove.seg}{(logical) Should segmentation from FAST be removed?}

\item{...}{additional arguments passed to \code{\link{fast}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function wraps a call to \code{fast} that performs bias
corretion
}

