% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_bin_tab.R
\name{fsl_bin_tab}
\alias{fsl_bin_tab}
\title{Quick Tabulation for logical images}
\usage{
fsl_bin_tab(x, y, dnames = c("x", "y"), verbose = FALSE)
}
\arguments{
\item{x}{filename of logical or 0/1 image}

\item{y}{filename of logical or 0/1 vimage}

\item{dnames}{names for table}

\item{verbose}{Should fsl commands be printed?}
}
\value{
table of x vs y
}
\description{
Creates a 2 by 2 table for
}
\note{
\code{fsl_bin} will be run to make these images binary before running
}
