% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_data_dir.R
\name{fsl_data_dir}
\alias{fsl_data_dir}
\title{Get FSL's Data Directory}
\usage{
fsl_data_dir()
}
\value{
Character path
}
\description{
Finds the FSLDIR from system environment or \code{getOption("fsl.path")}
and pastes on ``data''
}

