% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_standard_space_roi.R
\name{fsl_standard_space_roi}
\alias{fsl_standard_space_roi}
\title{FSL Standard Space ROI}
\usage{
fsl_standard_space_roi(
  file,
  outfile = NULL,
  mask = "FOV",
  roi = "FOV",
  retimg = TRUE,
  reorient = FALSE,
  intern = FALSE,
  opts = "",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{outfile}{(character) resultant image name (optional)}

\item{mask}{mask output using transformed standard space `"FOV"` (default),
`"NONE"` (no mask), or give a mask NIfTI file.}

\item{roi}{cut down input FOV using bounding box of the transformed standard 
space `"FOV"` (default), `"NONE"` (no mask), or give a mask NIfTI file.}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link[neurobase]{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fast}}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link[neurobase]{readnii}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This masks the input and/or reduces its FOV, on the basis of a
standard space image or mask, that is transformed into the space
of the input image.
}
