% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslabs.help.R
\name{fslabs.help}
\alias{fslabs.help}
\title{fslabs Help}
\usage{
fslabs.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fslabs} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslabs.help() 
}
}

