% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslacos.R
\name{fslacos.help}
\alias{fslacos.help}
\title{fslacos Help}
\usage{
fslacos.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslacos}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslacos.help()
}
}

