% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fsladd.R
\name{fsladd.help}
\alias{fsladd.help}
\title{fsladd Help}
\usage{
fsladd.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fsladd}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fsladd.help()
}
}

