% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsladd.R
\name{fsladd}
\alias{fsladd}
\title{Add Images using FSL}
\usage{
fsladd(file, file2, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", ...)
}
\arguments{
\item{file}{(character) input image}

\item{file2}{(character) image to be added}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented
when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fslmaths}}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslmaths -add}.
The R functions wraps \code{fslmaths}
}

