% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsland.R
\name{fsland}
\alias{fsland}
\alias{fsl_and}
\title{Logical AND with Images using FSL}
\usage{
fsland(file, file2, ...)

fsl_and(file, file2, ...)
}
\arguments{
\item{file}{(character) input image}

\item{file2}{(character) image to be multiplied}

\item{...}{additional arguments passed to \code{\link{fslmul}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function multiplies two images 
using \code{\link{fslmul}}) after
binarizing the images (using \code{\link{fslbin}}
}
\note{
Functions with underscores have different defaults
and will return an output filename, so to be used for piping
}
