% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslasin.help.R
\name{fslasin.help}
\alias{fslasin.help}
\title{fslasin Help}
\usage{
fslasin.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fslasin} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslasin.help() 
}
}

