% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslatan.R
\name{fslatan.help}
\alias{fslatan.help}
\title{fslatan Help}
\usage{
fslatan.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslatan}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslatan.help()
}
}

