% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslbet}
\alias{fslbet}
\title{Use FSL's Brain Extraction Tool (BET)}
\usage{
fslbet(infile, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", betcmd = c("bet2", "bet"), verbose = TRUE,
  ...)
}
\arguments{
\item{infile}{(character) input filename}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{opts}{(character) additional options to \code{bet}}

\item{betcmd}{(character) Use \code{bet} or \code{bet2} function}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
character or logical depending on intern
}
\description{
This function calls \code{bet} to extract a brain 
from an image, usually for skull stripping.
}

