% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslbin.help.R
\name{fslbin.help}
\alias{fslbin.help}
\title{fslbin Help}
\usage{
fslbin.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslbin} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslbin.help()
}
}

