% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslbinv.help.R
\name{fslbinv.help}
\alias{fslbinv.help}
\title{fslbinv Help}
\usage{
fslbinv.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslbinv} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslbinv.help()
}
}

