% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslchfiletype.R
\name{fslchfiletype.help}
\alias{fslchfiletype.help}
\title{fslchfiletype help}
\usage{
fslchfiletype.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslchfiletype}'s help
}
\examples{
if (have.fsl()){
 fslchfiletype.help()
}  
}
