% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslchfiletype.R
\name{fslchfiletype}
\alias{fslchfiletype}
\title{FSL Change file type}
\usage{
fslchfiletype(file, filetype = "NIFTI_GZ", outfile = NULL,
  retimg = TRUE, reorient = FALSE, intern = FALSE, verbose = TRUE,
  ...)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{filetype}{filetype to change image to}

\item{outfile}{Output filename. If NULL, will overwrite input file}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If \code{retimg}, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslchfiletype}
}
