% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslcmd}
\alias{fslcmd}
\title{FSL Command Wrapper}
\usage{
fslcmd(func, file, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", verbose = TRUE, samefile = FALSE,
  opts_after_outfile = FALSE, frontopts = "", no.outfile = FALSE,
  trim_front = FALSE, run = TRUE, ...)
}
\arguments{
\item{func}{(character) FSL function}

\item{file}{(character) image to be manipulated}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{func}}

\item{verbose}{(logical) print out command before running}

\item{samefile}{(logical) is the output the same file?}

\item{opts_after_outfile}{(logical) should \code{opts} come after 
the \code{outfile} in the FSL command?}

\item{frontopts}{(character) options/character to put in before filename}

\item{no.outfile}{(logical) is there an output file in the arguments of 
the FSL function?}

\item{trim_front}{trim the whitespace from the front of the command.}

\item{run}{(logical) Should the command just be printed (if \code{FALSE})?}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls fsl command passed to \code{func}
}
