% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslcpgeom.R
\name{fslcpgeom.help}
\alias{fslcpgeom.help}
\title{fslcpgeom help}
\usage{
fslcpgeom.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslcpgeom}'s help
}
\examples{
if (have.fsl()){
 fslcpgeom.help()
}
}

