% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslcpgeom.R
\name{fslcpgeom}
\alias{fslcpgeom}
\title{FSL Copy Geometry}
\usage{
fslcpgeom(file, file_with_header, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{file_with_header}{image with header to be copied over}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fslmaths}}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslcpgeom}
}

