% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fsldir}
\alias{fsldir}
\title{Get FSL's Directory}
\usage{
fsldir()
}
\value{
Character path
}
\description{
Finds the FSLDIR from system environment or \code{getOption("fsl.path")}
for location of FSL fuctions and returns it
}

