% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsldiv.help.R
\name{fsldiv.help}
\alias{fsldiv.help}
\title{fsldiv Help}
\usage{
fsldiv.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fsldiv} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fsldiv.help() 
}
}

