% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslentropy.R
\name{fslentropy}
\alias{fslentropy}
\title{Image Mean Entropy}
\usage{
fslentropy(img, nonzero = FALSE, verbose = TRUE, ts = FALSE)
}
\arguments{
\item{img}{Object of class nifti, or path of file}

\item{nonzero}{(logical) Should the statistic be taken over non-zero voxels}

\item{verbose}{(logical) print out command before running}

\item{ts}{(logical) is the series a timeseries (4D), invoking \code{-t} 
option}
}
\value{
Vector of unless ts option invoked, then matrix
}
\description{
Estimates Mean Entropy of Image from FSL
}
\note{
This uses option -e or -E in \code{\link{fslstats}}
}

