% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslerode.help.R
\name{fslerode.help}
\alias{fslerode.help}
\title{fslerode Help}
\usage{
fslerode.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslerode} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslerode.help()
}
}

