% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslexp.help.R
\name{fslexp.help}
\alias{fslexp.help}
\title{fslexp Help}
\usage{
fslexp.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fslexp} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslexp.help() 
}
}

