% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_exp.R, R/fslexp.R
\name{fsl_exp}
\alias{fsl_exp}
\alias{fslexp}
\title{Exponentiate Image using FSL}
\usage{
fsl_exp(..., outfile = tempfile(fileext = ".nii.gz"), retimg = FALSE)

fslexp(file, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", ...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link{readnii}}.}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{file}{(character) input image to exponentiated}

\item{reorient}{(logical) If retimg, should file be reoriented 
when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fslmaths}}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslmaths -exp}.  
The R functions wraps \code{fslmaths}
}
\note{
Functions with underscores have different defaults
and will return an output filename, so to be used for piping
}
