% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fslfill2}
\alias{fslfill2}
\title{Fill image holes with dilation then erosion}
\usage{
fslfill2(file, outfile = NULL, kopts = "", remove.ends = TRUE,
  retimg = FALSE, reorient = FALSE, intern = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) filename of image to be filled}

\item{outfile}{(character) name of resultant filled file}

\item{kopts}{(character) Options passed for kernel before erosion/dilation}

\item{remove.ends}{(logical) Remove top and bottom dilation.}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
character or logical depending on intern
}
\description{
This function calls \code{fslmaths} to dilate an image, then calls
it again to erode it.
}
\note{
This function binarizes the image before running.
}

