% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fslhd.help}
\alias{fslhd.help}
\title{FSLhd help}
\usage{
fslhd.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslhd}'s help
}
\examples{
if (have.fsl()){
 fslhd.help()
}
}

