% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fslhd.parse}
\alias{fslhd.parse}
\title{Parse FSL Header}
\usage{
fslhd.parse(hd)
}
\arguments{
\item{hd}{(character) header from \code{\link{fslhd}}}
}
\value{
data.frame of information from FSL header
}
\description{
This function takes in a FSL header and parses the components
}

