% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fslhd}
\alias{fslhd}
\title{Get NIfTI header using FSL}
\usage{
fslhd(file, opts = "")
}
\arguments{
\item{file}{(character) image to be masked}

\item{opts}{(character) additional options to be passed to fslhd}
}
\value{
Character of infromation from fslhd
}
\description{
This function calls \code{fslhd} to obtain a nifti header
}

