% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslhelp}
\alias{fslhelp}
\title{Wrapper for getting fsl help}
\usage{
fslhelp(func_name, help.arg = "--help", extra.args = "")
}
\arguments{
\item{func_name}{FSL function name}

\item{help.arg}{Argument to print help, usually "--help"}

\item{extra.args}{Extra arguments to be passed other than 
\code{--help}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function takes in the function and returns the
help from FSL for that function
}
