% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslindex.R
\name{fslindex.help}
\alias{fslindex.help}
\title{fslindex Help}
\usage{
fslindex.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslindex}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslindex.help()
}
}

