% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslmask.help.R
\name{fslmask.help}
\alias{fslmask.help}
\title{fslmask Help}
\usage{
fslmask.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslmask} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslmask.help()
}
}

