% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fslmaths.help}
\alias{fslerode.help}
\alias{fslfill.help}
\alias{fslmask.help}
\alias{fslmaths.help}
\alias{fslsmooth.help}
\alias{fslsub2.help}
\alias{fslthresh.help}
\title{FSL Maths Help}
\usage{
fslmaths.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help
}
\examples{
if (have.fsl()){
 fslmaths.help()
}
}

