% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fslmaths}
\alias{fslmaths}
\title{FSL Maths}
\usage{
fslmaths(file, outfile = NULL, retimg = FALSE, reorient = FALSE,
  intern = TRUE, opts = "", ...)
}
\arguments{
\item{file}{(character) image to be smoothed}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fslmaths}}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
If \code{retimg} then object of class nifit.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslmaths}
}

