% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fslmerge.help}
\alias{fslmerge.help}
\title{FSLMerge help}
\usage{
fslmerge.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmerge}'s help
}
\examples{
if (have.fsl()){
 fslmerge.help()
}
}

