% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslmul.help.R
\name{fslmul.help}
\alias{fslmul.help}
\title{fslmul Help}
\usage{
fslmul.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fslmul} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslmul.help() 
}
}
