% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_mul.R, R/fslmul.R
\name{fsl_mul}
\alias{fsl_mul}
\alias{fslmul}
\title{Multiply Images using FSL}
\usage{
fsl_mul(..., outfile = tempfile(fileext = ".nii.gz"), retimg = FALSE)

fslmul(file, file2, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", ...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link{readnii}}.}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{file}{(character) input image}

\item{file2}{(character) image to be multiplied}

\item{reorient}{(logical) If retimg, should file be reoriented 
when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fslmaths}}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslmaths -mul}.  
The R functions wraps \code{fslmaths}
}
\note{
Functions with underscores have different defaults
and will return an output filename, so to be used for piping
}
