% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslnan.help.R
\name{fslnan.help}
\alias{fslnan.help}
\title{fslnan Help}
\usage{
fslnan.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslnan} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslnan.help()
}
}

