% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslnanm.R
\name{fslnanm.help}
\alias{fslnanm.help}
\title{fslnanm Help}
\usage{
fslnanm.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslnanm}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslnanm.help()
}
}

