% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fslorient.help}
\alias{fslorient.help}
\title{fslorient help}
\usage{
fslorient.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslorient}'s help
}
\examples{
if (have.fsl()){
 fslorient.help()
}
}

