% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslgetorient.R
\name{fslgetorient}
\alias{fslgetorient}
\alias{fslgetqform}
\alias{fslgetqformcode}
\alias{fslgetsform}
\alias{fslgetsformcode}
\title{FSL Orientation Wrappers}
\usage{
fslgetorient(file, verbose = TRUE)

fslgetsform(file, verbose = TRUE)

fslgetqform(file, verbose = TRUE)

fslgetsformcode(file, verbose = TRUE)

fslgetqformcode(file, verbose = TRUE)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{verbose}{(logical) print out command before running}
}
\value{
Result from system command, output from FSL
}
\description{
This function calls \code{fslorient -get*} and is a 
simple wrapper of \code{\link{fslorient}}
}

