% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslgetorient.R
\name{fslorienter}
\alias{fslorienter}
\title{Wrapper for FSL Get Orientation}
\usage{
fslorienter(file, opts = "", verbose = TRUE)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{opts}{option to send to fslorient}

\item{verbose}{(logical) print out command before running}
}
\value{
Result from system command, output from FSL
}
\description{
This function calls \code{fslorient -getorient} and is a 
simple wrapper of \code{\link{fslorient}}
}
