% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslrandn.R
\name{fslrandn.help}
\alias{fslrandn.help}
\title{fslrandn Help}
\usage{
fslrandn.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslrandn}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslrandn.help()
}
}

