% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fslrange}
\alias{fslrange}
\title{Get range of an image}
\usage{
fslrange(file, verbose = TRUE)
}
\arguments{
\item{file}{(character) filename of image to be checked}

\item{verbose}{(logical) print out command before running}
}
\value{
numeric vector of length 2
}
\description{
This function calls \code{fslstats -R} to get the range of an image
}

