% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslrem.R
\name{fslrem.help}
\alias{fslrem.help}
\title{fslrem Help}
\usage{
fslrem.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslrem}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslrem.help()
}
}

