% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslreorient2std.help}
\alias{fslreorient2std.help}
\title{fslreorient2std help}
\usage{
fslreorient2std.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslreorient2std}'s help
}
\examples{
if (have.fsl()){
 fslreorient2std.help()
}  
}

