% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslrobustfov.R
\name{fslrobustfov}
\alias{fslrobustfov}
\alias{fsl_robustfov}
\title{FSL Robust Field of View}
\usage{
fslrobustfov(file, brain_size = NULL, mat_name = NULL,
  roi_name = NULL, retimg = TRUE, reorient = FALSE, intern = FALSE,
  verbose = TRUE, ...)

fsl_robustfov(retimg = FALSE, ...)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{brain_size}{size of brain in z-dimension (default 150mm)}

\item{mat_name}{matrix output name}

\item{roi_name}{ROI volume output name}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If \code{retimg}, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{robustfov} to
automatically crop the image
}
