% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslsin.R
\name{fslsin.help}
\alias{fslsin.help}
\title{fslsin Help}
\usage{
fslsin.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslsin}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslsin.help()
}
}

