% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslsplit.R
\name{fslsplit.help}
\alias{fslsplit.help}
\title{FSL Split help}
\usage{
fslsplit.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslsplit}'s help
}
\examples{
if (have.fsl()){
 fslsplit.help()
}  
}
