% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslsqr.R
\name{fslsqr.help}
\alias{fslsqr.help}
\title{fslsqr Help}
\usage{
fslsqr.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslsqr}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslsqr.help()
}
}

