% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslsqr.R
\name{fslsqr}
\alias{fslsqr}
\title{Square Image using FSL}
\usage{
fslsqr(file, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", ...)
}
\arguments{
\item{file}{(character) input image to square}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented
when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fslmaths}}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslmaths -sqr}.
The R functions wraps \code{fslmaths}
}

