% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslsqrt.R
\name{fslsqrt.help}
\alias{fslsqrt.help}
\title{fslsqrt Help}
\usage{
fslsqrt.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslsqrt}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslsqrt.help()
}
}

