% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslstats.help}
\alias{fslstats.help}
\alias{fslrange.help}
\alias{fslmean.help}
\alias{fslentropy.help}
\alias{fslsd.help}
\title{FSL Stats Help}
\usage{
fslstats.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslstats}'s help
}
\examples{
if (have.fsl()){
 fslstats.help() 
}
}
