% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslsub2.help.R
\name{fslsub2.help}
\alias{fslsub2.help}
\title{fslsub2 Help}
\usage{
fslsub2.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslsub2} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslsub2.help()
}
}

