\name{fslsub2}
\alias{fslsub2}
\title{Subsample image by factor of 2}
\usage{
fslsub2(file, outfile = file, intern = TRUE)
}
\arguments{
  \item{file}{(character) filename of image to be
  thresholded}

  \item{outfile}{(character) name of resultant subsampled
  file}

  \item{intern}{(logical) pass to \code{\link{system}}}
}
\value{
character or logical depending on intern
}
\description{
Subsample image by factor of 2
}

