% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslswapdim.help}
\alias{fslswapdim.help}
\title{fslswapdim help}
\usage{
fslswapdim.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslswapdim}'s help
}
\examples{
if (have.fsl()){
 fslswapdim.help()
}  
}
