% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fslswapdim}
\alias{fslswapdim}
\title{FSL Swap Dimensions}
\usage{
fslswapdim(file, outfile = NULL, retimg = FALSE, reorient = FALSE,
  intern = TRUE, a = "x", b = "y", c = "z", verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{a}{(character) Option for x domain in \code{fslswapdim}}

\item{b}{(character) Option for y domain in \code{fslswapdim}}

\item{c}{(character) Option for z domain in \code{fslswapdim}}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslswapdim}
}

