% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fsltan.R
\name{fsltan.help}
\alias{fsltan.help}
\title{fsltan Help}
\usage{
fsltan.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fsltan}is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fsltan.help()
}
}

