% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fslthresh.help.R
\name{fslthresh.help}
\alias{fslthresh.help}
\title{fslthresh Help}
\usage{
fslthresh.help(...)
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslthresh} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslthresh.help()
}
}

