% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslval.help}
\alias{fslval.help}
\title{fslval help}
\usage{
fslval.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslval}'s help
}
\examples{
if (have.fsl()){
 fslval.help()
} 
}

