% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslvol.R
\name{fslvol}
\alias{fslvol}
\title{FSL Volume in mL (or cubic centimeters)}
\usage{
fslvol(file, ...)
}
\arguments{
\item{file}{(character) filename of image to be checked}

\item{...}{options passed to \code{\link{fslsum}}}
}
\value{
Numeric value of volume in mL
}
\description{
This function wraps \code{\link{fslsum}} and 
\code{\link{voxdim}}
}
\note{
This may be approximate due to rounding
}

