% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Get FSLDir}
\alias{Get FSLDir}
\alias{fsldir}
\title{Get FSL's Directory}
\usage{
fsldir()
}
\value{
Character path
}
\description{
Finds the FSLDIR from system environment or \code{getOption("fsl.path")}
for location of FSL fuctions and returns it
}

