% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.fsloutput}
\alias{get.fsloutput}
\title{Determine FSL output type}
\usage{
get.fsloutput()
}
\value{
FSLOUTPUTTYPE, such as NIFTI_GZ.  If none found, uses NIFTI_GZ as default
}
\description{
Finds the FSLOUTPUTTYPE from system environment or
\code{getOption("fsl.outputtype")} for output type (nii.gz, nii, ANALYZE,etc)
}

