% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.imgext}
\alias{get.imgext}
\title{Determine extension of image based on FSLOUTPUTTYPE}
\usage{
get.imgext()
}
\value{
Extension for output type
}
\description{
Runs \code{get.fsloutput()} to extract FSLOUTPUTTYPE and then
gets corresponding extension (such as .nii.gz)
}

