% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmptyImageDimensions.R
\name{getEmptyImageDimensions}
\alias{getEmptyImageDimensions}
\alias{get_empty_dim}
\title{Get Empty Image Dimensions}
\usage{
getEmptyImageDimensions(img, value = 0, threshold = 0, reorient = FALSE)

get_empty_dim(img, value = 0, threshold = 0, reorient = FALSE)
}
\arguments{
\item{img}{nifti object}

\item{value}{Value to check against.  If zero, then 
\code{getEmptyImageDimensions} will include any dimension that has 
fewer than \code{threshold} zeroes. May be a vector of values, matched with 
\code{\link{match}}}

\item{threshold}{Include dimension if fewer than \code{threshold} voxels
are in the slice}

\item{reorient}{Should image be reoriented if a filename?}
}
\value{
List of length 3 of indices.
}
\description{
Creates a list of indices of an image that has all irrelevant
values
}
\note{
\code{get_empty_dim} is a shorthand for \code{getEmptyImageDimensions}
with all the same arguments.  Also, \code{NA} are set to zero.
}

