% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intent_code.R
\docType{methods}
\name{intent_code-methods}
\alias{intent_code,character-method}
\alias{intent_code-methods}
\title{Extract Image intent_code attribute}
\usage{
\S4method{intent_code}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
Extract Image intent_code attribute
}

