% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intent_p1.R
\docType{methods}
\name{intent_p1-methods}
\alias{intent_p1,character-method}
\alias{intent_p1-methods}
\title{Extract Image intent_p1 attribute}
\usage{
\S4method{intent_p1}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
Extract Image intent_p1 attribute
}

