% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intent_p2.R
\docType{methods}
\name{intent_p2-methods}
\alias{intent_p2,character-method}
\alias{intent_p2-methods}
\title{Extract Image intent_p2 attribute}
\usage{
\S4method{intent_p2}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
intent_p2 method for character types
}

