% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intent_p3.R
\docType{methods}
\name{intent_p3-methods}
\alias{intent_p3,character-method}
\alias{intent_p3-methods}
\title{Extract Image intent_p3 attribute}
\usage{
\S4method{intent_p3}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
Extract Image intent_p3 attribute
}

