% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Operators.R
\docType{methods}
\name{Logic}
\alias{&,nifti,nifti-method}
\alias{*,nifti,nifti-method}
\alias{+,nifti,nifti-method}
\alias{/,nifti,nifti-method}
\alias{Logic}
\alias{|,nifti,nifti-method}
\title{Logical Operators}
\usage{
\S4method{-}{nifti,nifti}(e1, e2)

\S4method{+}{nifti,nifti}(e1, e2)

\S4method{*}{nifti,nifti}(e1, e2)

\S4method{/}{nifti,nifti}(e1, e2)

\S4method{&}{nifti,nifti}(e1, e2)

\S4method{|}{nifti,nifti}(e1, e2)
}
\description{
Logical Operators
}

