% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\docType{methods}
\name{magic-methods}
\alias{magic,character-method}
\alias{magic-methods}
\title{Extract Image magic attribute}
\usage{
\S4method{magic}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
magic method for character types
}

