% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyEmptyImageDimensions.R, R/maskEmptyImageDimensions.R
\name{apply_empty_dim}
\alias{apply_empty_dim}
\alias{maskEmptyImageDimensions}
\title{Make Mask from Empty Image Dimensions}
\usage{
apply_empty_dim(img, inds, reorient = FALSE)

maskEmptyImageDimensions(img, ..., reorient = FALSE)
}
\arguments{
\item{img}{nifti object}

\item{reorient}{Should image be reoriented if a filename?}

\item{...}{Arguments to be passed to \code{\link{getEmptyImageDimensions}}.}
}
\value{
Object of class \code{nifti}
}
\description{
Make a mask of an image that has all irrelevant
values
}
\note{
\code{mask_empty_dim} is a shorthand for \code{maskEmptyImageDimensions}
with all the same arguments.
}
\seealso{
\code{\link{getEmptyImageDimensions}}
}

