% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcflirt.R
\name{mcflirt.help}
\alias{mcflirt.help}
\title{MCFLIRT help}
\usage{
mcflirt.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{mcflirt}'s help
}
\examples{
library(fslr)
if (have.fsl()){
 mcflirt.help()
} 
}
