% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcflirt.R
\name{mcflirt}
\alias{mcflirt}
\title{FSL Motion Correction}
\usage{
mcflirt(
  file,
  outfile = NULL,
  retimg = TRUE,
  reorient = FALSE,
  intern = FALSE,
  opts = "",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If \code{retimg}, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{mcflirt}.  Cannot use 
\code{-o} or \code{-verbose}, as output file should be specified in \code{outfile}.}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
it will have additional attributes in the \code{additional_files} field.
}
\description{
This function calls \code{mcflirt}
}
