% Generated by roxygen2 (4.0.2): do not edit by hand
\name{melodic.help}
\alias{melodic.help}
\title{MELODIC help}
\usage{
melodic.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{melodic}'s help
}
\examples{
if (have.fsl()){
 melodic.help()
}
}

