% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_fname.R
\name{mni_fname}
\alias{mni_fname}
\title{Construct MNI Filename}
\usage{
mni_fname(mm = c("1", "0.5", "2"), brain = FALSE, linear = FALSE,
  mask = FALSE)
}
\arguments{
\item{mm}{Resolution (in mm) of the brain image (isotropic)}

\item{brain}{Should the brain be returned (default) or the T1 with the skull}

\item{linear}{Should the linearized MNI template be used}

\item{mask}{should the mask be given?  Generally, only
MNI152_T1_1mm_brain_mask exists.}
}
\value{
Character path of filename, warning if that file does not exist
}
\description{
Finds the standard data directory for FSL and pastes togehter
the string for an MNI template image
}

