% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_fname.R
\name{mni_img}
\alias{mni_img}
\title{Read MNI Filename}
\usage{
mni_img(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{mni_fname}}}
}
\value{
Object of class \code{\link{nifti}}
}
\description{
Simple wrapper for reading in the MNI image constructed from 
\code{\link{mni_fname}}
}
