% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_img.R
\name{newnii}
\alias{newnii}
\title{Resets image parameters for a copied nifti object}
\usage{
newnii(img, ...)
}
\arguments{
\item{img}{nifti object (or character of filename)}

\item{...}{arguments to be passed to \code{\link{datatype}}}
}
\value{
object of type nifti
}
\description{
Resets the slots of a nifti object, usually because an image
was loaded, then copied and filled in with new data instead of making a 
nifti object from scratch.  Just a wrapper for smaller functions
}

