% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiarr.R
\name{niftiarr}
\alias{niftiarr}
\title{Make new nifti from array}
\usage{
niftiarr(img, arr)
}
\arguments{
\item{img}{object of class nifti}

\item{arr}{array to be passed in to .Data slot}
}
\value{
object of class nifti
}
\description{
Make new nifti object by passing in old nifti and array
}

