% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nii.stub}
\alias{nii.stub}
\title{Grab nii file stubname}
\usage{
nii.stub(x, bn = FALSE)
}
\arguments{
\item{x}{character vector of filenames ending in .nii or .nii.gz}

\item{bn}{Take \code{\link{basename}} of file?}
}
\description{
Quick helper function to strip of .nii or .nii.gz from filename
}

