% Generated by roxygen2 (4.0.2): do not edit by hand
\name{onefile}
\alias{onefile}
\title{Creates onefile specification}
\usage{
onefile(img)
}
\arguments{
\item{img}{nifti object}
}
\value{
Object of class nifti
}
\description{
Changes the magic and vox_offset slots to be consistent
with onefile option in \code{\link{writeNIfTI}}.  As of version 0.4.0,
\code{oro.nifti} did not support "ni1" magic outputs for output.
}

