% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qform.R, R/sform.R
\docType{methods}
\name{qform,character-method}
\alias{qform,character,character-method}
\alias{qform,character-method}
\alias{sform,character,character-method}
\alias{sform,character-method}
\title{Extract NIfTI 3D Image Orientation}
\usage{
\S4method{qform}{character}(object)

\S4method{sform}{character}(object)
}
\description{
Extract NIfTI 3D Image Orientation

Extract NIfTI 3D Image Orientation
}

