% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avscale.R
\name{parse_avscale}
\alias{parse_avscale}
\title{Parse output from \code{avscale}}
\usage{
parse_avscale(av_out)
}
\arguments{
\item{av_out}{output from \code{\link{fsl_avscale}}, character vector}
}
\value{
List of output values
}
\description{
This function parses the output from 
\code{\link{fsl_avscale}} into something more manageable
}
