% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pixdim.R
\docType{methods}
\name{pixdim-methods}
\alias{pixdim-methods}
\title{Extract Image pixdim attribute}
\usage{
\S4method{pixdim}{character}(object)
}
\arguments{
\item{object}{is an object of class \code{character}}
}
\description{
Methods that extract the ``pixdim'' slot from the
NIfTI/ANALYZE header.
}

