% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_read_xfm.R
\name{read_xfm}
\alias{read_xfm}
\title{Read FSL Transformation}
\usage{
read_xfm(file)
}
\arguments{
\item{file}{transformation file from \code{\link{flirt}},
usually ending in `.mat`}
}
\value{
A numeric matrix of numeric class
}
\description{
Read FSL Transformation
}
