% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readNIfTI2.R
\name{readNIfTI2}
\alias{readNIfTI2}
\alias{readnii}
\title{readNIfTI with default non-reoirientation}
\usage{
readNIfTI2(..., reorient = FALSE)

readnii(..., reorient = FALSE)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{readNIfTI}}}

\item{reorient}{Reorientation argument to pass to \code{\link{readNIfTI}}}
}
\value{
\code{nifti} object
}
\description{
This function calls the \code{\link{readNIfTI}} function from the
\code{oro.nifti} package, but sets the reorientation to \code{FALSE} by default
}

