% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_img.R
\name{remake_img}
\alias{remake_img}
\title{Remake Image from Vector}
\usage{
remake_img(vec, img, mask = NULL, warn = FALSE, ...)
}
\arguments{
\item{vec}{vector of values to be in resulting image}

\item{img}{object of class \code{\link{nifti}} to put vector into}

\item{mask}{binary array/ \code{\link{nifti}} object to denote where
vector values are to be.}

\item{warn}{Should a warning be issued if defaulting to FLOAT32?}

\item{...}{additional arguments passed to \code{\link{datatyper}}}
}
\value{
object of class  \code{\link{nifti}}
}
\description{
Wrapper function to take a vector of values and result in a 
\code{\link{nifti}} object
}
\seealso{
\code{\link{niftiarr}}
}

