% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_orient.R
\name{reverse_rpi_orient}
\alias{reverse_rpi_orient}
\title{Reverse Reorientation an Image to RPI orientation}
\usage{
reverse_rpi_orient(file, convention = c("NEUROLOGICAL", "RADIOLOGICAL"),
  orientation, verbose = TRUE)
}
\arguments{
\item{file}{Object of class \code{nifti} or character path}

\item{convention}{Convention of original image (usually from \code{\link{rpi_orient}})}

\item{orientation}{Vector of length 3 fromoriginal image 
(usually from \code{\link{rpi_orient}})}

\item{verbose}{print diagnostic messages}
}
\value{
Object of class \code{nifti}
}
\description{
This function uses \code{fslswapdim} to reorient an image
}

