% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_first_all.R
\name{run_first_all.help}
\alias{run_first_all.help}
\title{Run FIRST All Help}
\usage{
run_first_all.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{run_first_all}'s help
}
\examples{
library(fslr)

if (have.fsl()){
 run_first_all.help() 
}
}
