% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scl_inter.R
\docType{methods}
\name{scl_inter}
\alias{scl_inter}
\title{Extract NIfTI 3D Image scl_inter attribute}
\usage{
scl_inter(object)

\S4method{scl_inter}{nifti}(object)

scl_inter(object) <- value

\S4method{scl_inter}{nifti}(object) <- value

\S4method{scl_inter}{character}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to scl_inter}
}
\description{
Methods that act on the ``scl_inter'' in the NIfTI header.
}

