% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeof_hdr.R
\docType{methods}
\name{sizeof_hdr-methods}
\alias{sizeof_hdr-methods}
\alias{sizeof_hdr,character-method}
\title{Extract Image sizeof_hdr attribute}
\usage{
\S4method{sizeof_hdr}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
`sizeof_hdr` method for character types
}
