% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_code.R
\docType{methods}
\name{slice_code-methods}
\alias{slice_code,character-method}
\alias{slice_code-methods}
\title{Extract Image slice_code attribute}
\usage{
\S4method{slice_code}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
slice_code method for character types
}

