% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_duration.R
\docType{methods}
\name{slice_duration-methods}
\alias{slice_duration,character-method}
\alias{slice_duration-methods}
\title{Extract Image slice_duration attribute}
\usage{
\S4method{slice_duration}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
slice_duration method for character types
}

