% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_end.R
\docType{methods}
\name{slice_end-methods}
\alias{slice_end,character-method}
\alias{slice_end-methods}
\title{Extract Image slice_end attribute}
\usage{
\S4method{slice_end}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
slice_end method for character types
}

