% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slice_start.R
\docType{methods}
\name{slice_start-methods}
\alias{slice_start,character-method}
\alias{slice_start-methods}
\title{Extract Image slice_start attribute}
\usage{
\S4method{slice_start}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
Extract Image slice_start attribute
}

