% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susan.R
\name{susan.help}
\alias{susan.help}
\title{FSL SUSAN Help}
\usage{
susan.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{susan}'s help
}
\examples{
library(fslr)
if (have.fsl()){
 susan.help() 
}
}
