% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tempimg}
\alias{tempimg}
\title{Create temporary nii.gz file for FSL}
\usage{
tempimg(nim, gzipped = TRUE, checknan = TRUE, ...)
}
\arguments{
\item{nim}{object of class nifti}

\item{gzipped}{Should file be gzipped? Passed to \code{\link{writeNIfTI}}}

\item{checknan}{Check for NAs or NaNs}

\item{...}{Not used}
}
\value{
filename of output nii.gz
}
\description{
Takes in a object of class nifti, writes it to a temp file, appends
.nii.gz as \code{\link{writeNIfTI}} adds it.
}

