% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vox_offset.R
\docType{methods}
\name{vox_offset}
\alias{vox_offset}
\title{Extract NIfTI 3D Image vox_offset attribute}
\usage{
vox_offset(object)

\S4method{vox_offset}{nifti}(object)

vox_offset(object) <- value

\S4method{vox_offset}{nifti}(object) <- value

\S4method{vox_offset}{character}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to vox_offset}
}
\description{
Methods that act on the ``vox_offset'' in the NIfTI header.
}

