% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nii.stub.R
\name{voxdim}
\alias{voxdim}
\title{Gets Voxel Dimensions}
\usage{
voxdim(img)
}
\arguments{
\item{img}{nifti object}
}
\value{
Vector of length 3
}
\description{
Grabs the pixdim and takes the correct elements
}

