% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writenii.R
\name{writeNIfTI2}
\alias{writeNIfTI2}
\alias{writenii}
\title{writeNIfTI with default non-reoirientation}
\usage{
writeNIfTI2(nim, filename, dtype = FALSE, ...)

writenii(nim, filename, dtype = FALSE, ...)
}
\arguments{
\item{nim}{object of class \code{nifti}, passed to \code{\link{writeNIfTI}}}

\item{filename}{path to save the NIfTI file.  Suffix will be removed}

\item{dtype}{Should \code{\link{datatyper}} be run before writing?}

\item{...}{Additional arguments pased to \code{\link{writeNIfTI}}}
}
\value{
Nothing
}
\description{
This function calls the \code{\link{writeNIfTI}} function from the
\code{oro.nifti} package, but makes sure to remove \code{.nii} extension
}

