% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_pad.R
\name{zero_pad}
\alias{zero_pad}
\title{Zero pads an image}
\usage{
zero_pad(img, kdim, invert = FALSE)
}
\arguments{
\item{img}{Array or class nifti}

\item{kdim}{Dimensions of kernel}

\item{invert}{(logical) If \code{FALSE}, does zero padding.  If \code{TRUE}, 
reverses the process.}
}
\value{
Array
}
\description{
This function zero pads an image by a certain number of 
dimensions, usually for convolution
}
\examples{
kdim = c(3,3,5)
img = array(rnorm(512*512*36), dim = c(512, 512, 36))
pad = zero_pad(img, kdim)
back = zero_pad(pad, kdim, invert=TRUE)
all.equal(back, img)
}

