% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zero_trans}
\alias{zero_trans}
\title{Change intercept to 0 and slope to 1}
\usage{
zero_trans(img)
}
\arguments{
\item{img}{nifti object (or character of filename)}
}
\value{
object of type nifti
}
\description{
Forces image scl_slope to 1 and scl_inter  to be 0 of slots of class
nifti.  This is so that when images are rendered/written, the values correspond to those
in the array (stored in Data slot) and are not scaled.
}

