% Generated by roxygen2 (4.0.1): do not edit by hand
\name{zscore_img}
\alias{zscore_img}
\title{Get Z-score over a margin of an img}
\usage{
zscore_img(img, margin = 3)
}
\arguments{
\item{img}{character path of image or
an object of class nifti}

\item{margin}{Margin of image to z-score over (3-Axial, 2-Sagittal,
1-Coronal)}
}
\value{
Array of object of class nifti
}
\description{
Standardizes an image either by the axial, sagittal, or
coronal slice
}
\examples{
dim = c(100, 30, 5)
img = array(rnorm(prod(dim), mean=4, sd=4),
dim=dim)

truth2 = img
for (i in 1:dim(img)[2]) {
truth2[,i,] = (truth2[,i,]- mean(truth2[,i,]))/sd(truth2[,i,])
}

truth1 = img
for (i in 1:dim(img)[1]) {
truth1[i,,] = (truth1[i,,]- mean(truth1[i,,]))/sd(truth1[i,,])
}

truth3 = img
for (i in 1:dim(img)[3]) {
truth3[,,i] = (truth3[,,i]- mean(truth3[,,i]))/sd(truth3[,,i])
}
try3 = zscore_img(img, margin=3)
print(all.equal(try3, truth3))
try2 = zscore_img(img, margin=2)
print(all.equal(try2, truth2))
try1 = zscore_img(img, margin=1)
print(all.equal(try1, truth1))
}
\seealso{
\link{aperm}
}

