% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitpix.R
\docType{methods}
\name{bitpix-methods}
\alias{bitpix-methods}
\alias{bitpix,character-method}
\title{Extract Image bitpix attribute}
\usage{
\S4method{bitpix}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
bitpix method for character types
}
