% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cal_img}
\alias{cal_img}
\title{Set Max/Min for nifti object by range of data}
\usage{
cal_img(img)
}
\arguments{
\item{img}{nifti object}
}
\value{
object of type nifti
}
\description{
Rescales image cal_max and cal_min slots to be the max and min,
respectively, of an object of class nifti, with \code{na.rm=TRUE}. This is so that
when images are rendered/written, the values correspond to those
in the array (stored in .Data slot) are plotted on correct grayscale and no error is
given by \code{writeNIfTI}.
}

