% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_max.R
\docType{methods}
\name{cal.max-methods}
\alias{cal.max-methods}
\alias{cal.max,character-method}
\title{Extract Image cal.max attribute}
\usage{
\S4method{cal.max}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
cal_max method for character types
}
