% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_min.R
\docType{methods}
\name{cal.min-methods}
\alias{cal.min-methods}
\alias{cal.min,character-method}
\title{Extract Image cal.min attribute}
\usage{
\S4method{cal.min}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
cal_min method for character types
}
