% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_file}
\alias{check_file}
\title{Wrapper for getForms with filename}
\usage{
check_file(file)
}
\arguments{
\item{file}{(character) filename of image to be checked}
}
\value{
result of \code{\link{checkout}}
}
\description{
Wrapper for getForms with filename
}

