% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslhd.R
\name{check_file}
\alias{check_file}
\title{Wrapper for getForms with filename}
\usage{
check_file(file, ...)
}
\arguments{
\item{file}{(character) filename of image to be checked}

\item{...}{options passed to \code{\link{checkimg}}}
}
\value{
result of \code{\link{checkout}}
}
\description{
Wrapper for getForms with filename
}
\examples{
if (have.fsl()){
 mnifile = file.path(fsldir(), "data", "standard",
   "MNI152_T1_2mm.nii.gz")
 check_file(mnifile)
}
}

