% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{check_file}
\alias{check_file}
\title{Wrapper for getForms with filename}
\usage{
check_file(file, ...)
}
\arguments{
\item{file}{(character) filename of image to be checked}

\item{...}{options passed to \code{\link{checkimg}}}
}
\value{
result of \code{\link{checkout}}
}
\description{
Checking the q/s-forms for a header
}
\examples{
library(fslr)
if (have.fsl()){
 mnifile = mni_fname("2")
 check_file(mnifile)
} 
}
