% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkimg}
\alias{checkimg}
\title{Check if filename is character or nifti object}
\usage{
checkimg(file)
}
\arguments{
\item{file}{character or nifti object}
}
\value{
character filename or temporary nii
}
\description{
Check if filename is character or nifti object
}

