% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkimg}
\alias{checkimg}
\title{Check if filename is character or nifti object}
\usage{
checkimg(file, ...)
}
\arguments{
\item{file}{character or nifti object}

\item{...}{options passed to \code{\link{tempimg}}}
}
\value{
character filename or temporary nii
}
\description{
Check if filename is character or nifti object
}

