% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checknii.R
\docType{methods}
\name{checknii-methods}
\alias{checknii}
\alias{checknii,character-method}
\alias{checknii,list-method}
\alias{checknii,nifti-method}
\alias{checknii-methods}
\alias{ensure_nii}
\title{Force object to filename with .nii extension}
\usage{
checknii(file, ...)

\S4method{checknii}{nifti}(file, ...)

\S4method{checknii}{character}(file, ...)

\S4method{checknii}{list}(file, ...)

ensure_nii(file, ...)
}
\arguments{
\item{file}{character or \code{nifti} object}

\item{...}{options passed to \code{\link{checkimg}}}
}
\value{
character filename of image or temporary nii, 
with .nii extension
}
\description{
Ensures the output to be a character filename (or vector) from an input
image or \code{nifti}, but not gzipped and has \code{.nii} extension
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}

