% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkout}
\alias{checkout}
\title{Determine of Q and S forms are consistent}
\usage{
checkout(hd)
}
\arguments{
\item{hd}{(list) sforms from \code{\link{getForms}}}
}
\value{
logical indicating if sform and qform consistent
}
\description{
This function determines if the determinants of the sform and qform
have the same sign
}

