% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho2.R
\name{colorbar}
\alias{colorbar}
\title{Add a colorbar to an ortho2 object}
\usage{
colorbar(breaks, col, text.col = "white", labels = TRUE, maxleft = 0.95)
}
\arguments{
\item{breaks}{a set of finite numeric breakpoints for the 
colours (see \code{\link{image}}}

\item{col}{a list of colors (see \code{\link{image}}}

\item{text.col}{axis and text label color}

\item{labels}{labels for tick marks - see \code{\link{axis}}}

\item{maxleft}{Extent the lefthand for colorbar}
}
\value{
A plot
}
\description{
Adds a series of colors mapped to a value
}
\note{
Much of this was taken from \code{vertical.image.legend} from
the \code{aqfig} package
}

