% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datatype.R
\docType{methods}
\name{datatype-methods}
\alias{datatype,character-method}
\alias{datatype-methods}
\title{Extract Image datatype attribute}
\usage{
\S4method{datatype}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
Extract Image datatype attribute
}

