% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_img.R
\name{datatype}
\alias{datatype}
\alias{datatyper}
\title{Change Data type for img}
\usage{
datatyper(img, type_string = NULL, datatype = NULL, bitpix = NULL,
  trybyte = TRUE, warn = TRUE)
}
\arguments{
\item{img}{nifti object (or character of filename)}

\item{type_string}{(NULL) character of datatype and bitpix.  Supercedes
both datatype and bitpix.  If specified 
\code{convert.datatype[[type_string]]} and 
\code{convert.bitpix[[type_string]]} will be used.}

\item{datatype}{(NULL) character of datatype see 
\code{\link{convert.datatype}}}

\item{bitpix}{(NULL) character of bitpix see 
\code{\link{convert.bitpix}}}

\item{trybyte}{(logical) Should you try to make a byte (UINT8) if image in
c(0, 1)?}

\item{warn}{Should a warning be issued if defaulting to FLOAT32?}
}
\value{
object of type nifti
}
\description{
Tries to figure out the correct datatype for image.  Useful 
for image masks - makes them binary if
}

