% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip.R
\docType{methods}
\name{descrip-methods}
\alias{descrip-methods}
\alias{descrip,character-method}
\title{Extract Image descrip attribute}
\usage{
\S4method{descrip}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
descrip method for character types
}
