% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dim_.R
\docType{methods}
\name{dim_-methods}
\alias{dim_,character-method}
\alias{dim_-methods}
\title{Extract Image dim_ attribute}
\usage{
\S4method{dim_}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
Extract Image dim_ attribute
}

