% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_.R
\docType{methods}
\name{dim_-methods}
\alias{dim_-methods}
\alias{dim_,character-method}
\title{Extract Image dim_ attribute}
\usage{
\S4method{dim_}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
\code{dim_} method for class character
}
