% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dropEmptyImageDimensions.R
\name{dropEmptyImageDimensions}
\alias{dropEmptyImageDimensions}
\title{Drop Empty Image Dimensions}
\usage{
dropEmptyImageDimensions(img, value = 0, threshold = 0, other.imgs = NULL,
  reorient = FALSE)
}
\arguments{
\item{img}{nifti object}

\item{value}{Value to check against.  If zero, then
\code{dropEmptyImageDimensions} will drop any dimension that has
all zeroes}

\item{threshold}{Drop dimension if fewer than \code{threshold} voxels
are in the slice}

\item{other.imgs}{List of other nifti objects or filenames
to apply the same dropping as \code{img}.}

\item{reorient}{Should image be reoriented if a filename?}
}
\value{
Object of class nifti
}
\description{
Drops dimensions of an image that has all irrelevant
values
}

