% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fast.R
\name{fast}
\alias{fast}
\title{FSL FAST}
\usage{
fast(file, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", out_type = c("seg", "mixeltype", "pve_0",
  "pve_1", "pve_2", "pveseg"), verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fast}}

\item{out_type}{(character) Suffix to grab from outfile.  For
example, output filename is \code{paste0(outfile, "_", out_type)}}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fast} from FSL
}

