% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{flirt.help}
\alias{flirt.help}
\alias{flirt_apply.help}
\title{FLIRT help}
\usage{
flirt.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{flirt}'s help
}
\examples{
if (have.fsl()){
 flirt.help()
} 
}
