% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnirt.R
\name{fnirt_with_affine_apply}
\alias{fnirt_with_affine_apply}
\title{Applies FLIRT then FNIRT transformations}
\usage{
fnirt_with_affine_apply(infile, reffile, flirt.omat = NULL,
  flirt.outfile = NULL, fnirt.warpfile = NULL, outfile = NULL,
  retimg = TRUE, reorient = FALSE, intern = FALSE, flirt.opts = "",
  opts = "", verbose = TRUE, ...)
}
\arguments{
\item{infile}{(character) input filename}

\item{reffile}{(character) reference image to be registered to}

\item{flirt.omat}{(character) Filename of output affine matrix}

\item{flirt.outfile}{(character) Filename of output affine-registered image}

\item{fnirt.warpfile}{(character) Filename of warp image from 
\code{\link{fnirt}}}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{flirt.opts}{(character) additional options to FLIRT}

\item{opts}{(character) additional options to FNIRT}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
character or logical depending on intern
}
\description{
Applies an affine transformation with FLIRT then the warp image with FNIRT
}
\seealso{
fnirt_with_affine
}

