% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_anat.R
\name{fsl_anat}
\alias{fsl_anat}
\title{FSL Anatomical Processing Script}
\usage{
fsl_anat(
  file,
  modality = c("T1", "T2", "PD"),
  outdir = NULL,
  intern = FALSE,
  opts = "",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{file}{(character) image to be manipulated, should be full path}

\item{modality}{(character) Modality of Image to be run}

\item{outdir}{(character) output directory, if none specified, will 
default to \code{dirname(file)}}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fsl_anat}}

\item{verbose}{(logical) print out command before running}

\item{...}{options passed to \code{\link{checkimg}}}
}
\value{
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fsl_anat} from FSL
}
