% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{fsl_biascorrect}
\alias{fsl_biascorrect}
\title{FSL Bias Correct}
\usage{
fsl_biascorrect(file, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", verbose = TRUE, remove.seg = TRUE, ...)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fast}}

\item{verbose}{(logical) print out command before running}

\item{remove.seg}{(logical) Should segmentation from FAST be removed?}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function wraps a call to \code{fast} that performs bias
corretion
}
