% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_dice.R
\name{fsl_dice}
\alias{fsl_dice}
\title{Calculate Dice Coefficient of 2 Binary images}
\usage{
fsl_dice(x, y, ...)
}
\arguments{
\item{x}{filename of logical or 0/1 image}

\item{y}{filename of logical or 0/1 image}

\item{...}{arguments passed to \code{\link{fsl_bin_tab}}}
}
\value{
Single number of the dice coefficient
}
\description{
Creates a 2 by 2 table for
}
