% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_version.R
\name{fsl_version}
\alias{fsl_version}
\alias{fslversion}
\alias{fsl_version_gt5}
\title{Find FSL Version}
\usage{
fsl_version()

fslversion()

fsl_version_gt5()
}
\value{
If the version file does not exist, it will throw a warning, but 
it will return an empty string.  Otherwise it will be a string of the version.
}
\description{
Finds the FSL version from \code{FSLDIR/etc/fslversion}
}
\note{
This will use \code{fsldir()} to get the directory
}
\examples{
if (have_fsl()) {
 fslversion()
 fsl_version()
}
}
