% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsladd.help.R
\name{fsladd.help}
\alias{fsladd.help}
\title{fsladd Help}
\usage{
fsladd.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fsladd} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fsladd.help() 
}
}
