% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fslbet.help}
\alias{fslbet.help}
\title{Help for FSL BET}
\usage{
fslbet.help(betcmd = c("bet2", "bet"))
}
\arguments{
\item{betcmd}{(character) Get help for \code{bet} or \code{bet2} function}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{bet}'s help
}
\examples{
if (have.fsl()){
 fslbet.help()
 fslbet.help("bet")
}
}

