% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslcmd}
\alias{fslcmd}
\title{FSL Command Wrapper}
\usage{
fslcmd(func, file, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", verbose = TRUE, samefile = FALSE, ...)
}
\arguments{
\item{func}{(character) FSL function}

\item{file}{(character) image to be manipulated}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{func}}

\item{verbose}{(logical) print out command before running}

\item{samefile}{(logical) is the output the same file?}

\item{...}{additional arguments passed to \code{\link{readNIfTI}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls fsl command passed to \code{func}
}

