% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslcos.help.R
\name{fslcos.help}
\alias{fslcos.help}
\title{fslcos Help}
\usage{
fslcos.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslcos} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslcos.help()
}
}

