% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fsldir}
\alias{fsl_dir}
\alias{fsldir}
\title{Get FSL's Directory}
\usage{
fsldir()

fsl_dir()
}
\value{
Character path
}
\description{
Finds the FSLDIR from system environment or \code{getOption("fsl.path")}
for location of FSL functions and returns it
}

