% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsldiv.R
\name{fsldiv}
\alias{fsldiv}
\title{Divide Images using FSL}
\usage{
fsldiv(file, file2, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", ...)
}
\arguments{
\item{file}{(character) input image}

\item{file2}{(character) image to be divided}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented 
when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fslmaths}}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslmaths -div}.  
The R functions wraps \code{fslmaths}
}

