% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsledge.help.R
\name{fsledge.help}
\alias{fsledge.help}
\title{fsledge Help}
\usage{
fsledge.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fsledge} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fsledge.help()
}
}

