% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslfill2.R
\name{fslfill2}
\alias{fslfill2}
\title{Fill image holes with dilation then erosion}
\usage{
fslfill2(
  file,
  outfile = NULL,
  kopts = "",
  remove.ends = TRUE,
  refill = TRUE,
  retimg = TRUE,
  reorient = FALSE,
  intern = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{file}{(character) filename of image to be filled}

\item{outfile}{(character) name of resultant filled file}

\item{kopts}{(character) Options passed for kernel before erosion/dilation}

\item{remove.ends}{(logical) Remove top and bottom dilation.}

\item{refill}{(logical) Run \code{\link{fslfill}} after dilation/erosion.}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
character or logical depending on intern
}
\description{
This function calls \code{fslmaths} to dilate an image, then calls
it again to erode it.
}
\note{
This function binarizes the image before running.
}
