% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslhd.help}
\alias{fslhd.help}
\title{FSLhd help}
\usage{
fslhd.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslhd}'s help
}
\examples{
if (have.fsl()){
 fslhd.help()
}
}

