% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslhd.parse}
\alias{fslhd.parse}
\title{Parse FSL Header}
\usage{
fslhd.parse(hd)
}
\arguments{
\item{hd}{(character) header from \code{\link{fslhd}}}
}
\value{
data.frame of information from FSL header
}
\description{
This function takes in a FSL header and parses the components
}
\examples{
if (have.fsl()){
 mnifile = file.path(fsldir(), "data", "standard", 
   "MNI152_T1_2mm.nii.gz")
 hd = fslhd(mnifile)
 fslhd.parse(hd)
}  
}

