% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslhd}
\alias{fslhd}
\title{Get NIfTI header using FSL}
\usage{
fslhd(file, opts = "", verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) image filename or character of class nifti}

\item{opts}{(character) additional options to be passed to fslhd}

\item{verbose}{(logical) print out command before running}

\item{...}{options passed to \code{\link{checkimg}}}
}
\value{
Character of infromation from fslhd
}
\description{
This function calls \code{fslhd} to obtain a nifti header
}
\examples{
if (have.fsl()){
 mnifile = file.path(fsldir(), "data", "standard", 
   "MNI152_T1_2mm.nii.gz")
 fslhd(mnifile)
}   
}

