% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslmaths.help}
\alias{fslmaths.help}
\title{FSL Maths Help}
\usage{
fslmaths.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help
}
\examples{
if (have.fsl()){
 fslmaths.help()
}
}

