% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslmerge.help}
\alias{fslmerge.help}
\title{FSLMerge help}
\usage{
fslmerge.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmerge}'s help
}
\examples{
if (have.fsl()){
 fslmerge.help()
}  
}
