% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R, R/fslsplit.R
\name{fslmerge}
\alias{fslmerge}
\title{Merge images using FSL}
\usage{
fslmerge(infile, direction = c("t", "x", "y", "z"), output_basename = NULL,
  retimg = TRUE, reorient = FALSE, verbose = TRUE)

fslmerge(infile, direction = c("t", "x", "y", "z"), output_basename = NULL,
  retimg = TRUE, reorient = FALSE, verbose = TRUE)
}
\arguments{
\item{infile}{(character) input filename}

\item{direction}{(character) direction to merge over, x, y, z, 
t (time), a (auto)}

\item{output_basename}{(character) prefix to have for output}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{verbose}{(logical) print out command before running}

\item{infiles}{(character) input filenames}

\item{outfile}{(character) output filename}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{...}{additional arguments passed to \code{\link{readnii}}.}

\item{direction}{(character) direction to split over: t (time), x, y, z}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{verbose}{(logical) print out command before running}
}
\value{
character or logical depending on intern

List of output files
}
\description{
This function calls \code{fslmerge} to merge files on some dimension
and either saves the image or returns an object of class nifti

This function calls \code{fslsplit} to merge files on some dimension
and either saves the image or returns an object of class nifti
}

