% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslnanm.help.R
\name{fslnanm.help}
\alias{fslnanm.help}
\title{fslnanm Help}
\usage{
fslnanm.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fslnanm} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslnanm.help() 
}
}

