% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslorient}
\alias{fslorient}
\title{FSL Orient}
\usage{
fslorient(file, retimg = TRUE, reorient = FALSE, intern = FALSE,
  opts = "", verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If \code{retimg}, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{intern}{(logical) to be passed to \code{\link{system}}}

\item{opts}{(character) operations to be passed to \code{fslorient}}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{readnii}}.}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function calls \code{fslorient}
}
