% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslrand.help.R
\name{fslrand.help}
\alias{fslrand.help}
\title{fslrand Help}
\usage{
fslrand.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fslrand} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslrand.help() 
}
}
