% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslrecip.help.R
\name{fslrecip.help}
\alias{fslrecip.help}
\title{fslrecip Help}
\usage{
fslrecip.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslrecip} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslrecip.help()
}
}

