% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslsin.help.R
\name{fslsin.help}
\alias{fslsin.help}
\title{fslsin Help}
\usage{
fslsin.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fslsin} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslsin.help() 
}
}
