% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslsmooth.help.R
\name{fslsmooth.help}
\alias{fslsmooth.help}
\title{fslsmooth Help}
\usage{
fslsmooth.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslsmooth} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslsmooth.help()
}
}

