% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslsqr.help.R
\name{fslsqr.help}
\alias{fslsqr.help}
\title{fslsqr Help}
\usage{
fslsqr.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as
\code{fslsqr} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslsqr.help()
}
}

