% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslsqrt.help.R
\name{fslsqrt.help}
\alias{fslsqrt.help}
\title{fslsqrt Help}
\usage{
fslsqrt.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fslsqrt} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslsqrt.help() 
}
}

