% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fslstats.help}
\alias{fslrange.help}
\alias{fslstats.help}
\title{FSL Stats Help}
\usage{
fslstats.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslstats}'s help
}
\examples{
if (have.fsl()){
 fslstats.help()
}
}

