% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fslswapdim.help}
\alias{fslswapdim.help}
\title{fslswapdim help}
\usage{
fslswapdim.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslswapdim}'s help
}
\examples{
if (have.fsl()){
 fslswapdim.help()
}
}

