% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslthresh.help.R
\name{fslthresh.help}
\alias{fslthresh.help}
\title{fslthresh Help}
\usage{
fslthresh.help(...)
}
\arguments{
\item{...}{passed to \code{\link{fslmaths.help}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslmaths}'s help, as 
\code{fslthresh} is a wrapper for \code{fslmaths}
}
\examples{
if (have.fsl()){
 fslthresh.help() 
}
}

