% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fslval}
\alias{fslval}
\title{Get value from FSL header}
\usage{
fslval(file, keyword = "")
}
\arguments{
\item{file}{(character) image filename or character of class nifti}

\item{keyword}{(character) keyword to be taken from fslhd}
}
\value{
Character of infromation from fslhd field specified in keyword
}
\description{
This function calls \code{fslval} to obtain a nifti header
}

