% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fslview.help}
\alias{fslview.help}
\title{FSLView help}
\usage{
fslview.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{fslview}'s help
}
\examples{
if (have.fsl()){
 fslview.help()
}
}

