% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslhd.R
\name{fslview}
\alias{fslview}
\title{Open image in FSLView}
\usage{
fslview(file, intern = TRUE, opts = "", verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) filename of image to be thresholded}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{opts}{(character) options for FSLView}

\item{verbose}{(logical) print out command before running}

\item{...}{options passed to \code{\link{checkimg}}}
}
\value{
character or logical depending on intern
}
\description{
This function calls \code{fslview} to view an image
in the FSL viewer
}

