% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fslhd.R
\name{get.fsl}
\alias{get.fsl}
\title{Create command declaring FSLDIR}
\usage{
get.fsl()
}
\value{
NULL if FSL in path, or bash code for setting up FSL DIR
}
\description{
Finds the FSLDIR from system environment or \code{getOption("fsl.path")}
for location of FSL fuctions
}

