% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{get.imgext}
\alias{get.imgext}
\title{Determine extension of image based on FSLOUTPUTTYPE}
\usage{
get.imgext()
}
\value{
Extension for output type
}
\description{
Runs \code{get.fsloutput()} to extract FSLOUTPUTTYPE and then 
gets corresponding extension (such as .nii.gz)
}
