% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getForms}
\alias{getForms}
\title{Get Q and S Forms of orientation matrix}
\usage{
getForms(file)
}
\arguments{
\item{file}{(character) filename of image to pass to header}
}
\value{
list with elements of sform and qform and their respective codes
}
\description{
This function obtains the s and q forms of an image transformation
matrix
}

