% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{have.fsl}
\alias{have.fsl}
\alias{have_fsl}
\title{Logical check if FSL is accessible}
\usage{
have.fsl(...)

have_fsl(...)
}
\arguments{
\item{...}{options to pass to \code{\link{get.fsl}}}
}
\value{
Logical TRUE is FSL is accessible, FALSE if not
}
\description{
Uses \code{get.fsl} to check if FSLDIR is accessible or the option
\code{fsl.path} is set and returns logical
}
\examples{
have.fsl()
}
