% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intent_name.R
\docType{methods}
\name{intent_name-methods}
\alias{intent_name,character-method}
\alias{intent_name-methods}
\title{Extract Image intent_name attribute}
\usage{
\S4method{intent_name}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
intent_name method for character types
}

