% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intent_p1.R
\docType{methods}
\name{intent_p1-methods}
\alias{intent_p1-methods}
\alias{intent_p1,character-method}
\title{Extract Image intent_p1 attribute}
\usage{
\S4method{intent_p1}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
intent_p1 method for character types
}
