% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslhd.R
\name{melodic.help}
\alias{melodic.help}
\title{MELODIC help}
\usage{
melodic.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{melodic}'s help
}
\examples{
if (have.fsl()){
 melodic.help()
} 
}

