% Generated by roxygen2 (4.0.2): do not edit by hand
\name{melodic}
\alias{melodic}
\title{Run MELODIC ICA}
\usage{
melodic(file, outdir = dirname(file), intern = TRUE, opts = "",
  verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) image to be run}

\item{outdir}{(character) output directory.
(Default \code{dirname(file)})}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{opts}{(character) options for melodic}

\item{verbose}{(logical) print out command before running}

\item{...}{arguments passed to \code{\link{checkimg}}}
}
\value{
character or logical depending on intern
}
\description{
This function calls \code{melodic}
}

