% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qform.R, R/sform.R
\name{qform,character-method}
\alias{qform,character-method}
\alias{qform,character}
\alias{sform,character-method}
\alias{sform,character}
\title{Extract NIfTI 3D Image Orientation}
\usage{
\S4method{qform}{character}(object)

\S4method{sform}{character}(object)
}
\arguments{
\item{object}{is a \code{nifti} object}
}
\description{
Gets q/s-forms from a character
}
