% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cal_max.R
\docType{methods}
\name{pixdim,character-method}
\alias{pixdim,character,character-method}
\alias{pixdim,character-method}
\title{Extract or Replace NIfTI/Analyze Pixel Dimensions}
\usage{
\S4method{pixdim}{character}(object)
}
\description{
Extract or Replace NIfTI/Analyze Pixel Dimensions
}

