% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pixdim.R
\docType{methods}
\name{pixdim-methods}
\alias{pixdim,character-method}
\alias{pixdim-methods}
\title{Extract Image pixdim attribute}
\usage{
\S4method{pixdim}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
Extract Image pixdim attribute
}

