% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qform_code.R
\docType{methods}
\name{qform_code-methods}
\alias{qform_code-methods}
\alias{qform_code,character-method}
\title{Extract Image qform_code attribute}
\usage{
\S4method{qform_code}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
qform_code method for character types
}
