% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNIfTI2.R
\name{readNIfTI2}
\alias{readNIfTI2}
\alias{readnii}
\title{readNIfTI with default non-reoirientation}
\usage{
readNIfTI2(..., reorient = FALSE)

readnii(..., reorient = FALSE, dtype = TRUE, drop_dim = TRUE,
  warn = FALSE)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{readNIfTI}}}

\item{reorient}{Reorientation argument to pass to \code{\link{readNIfTI}}}

\item{dtype}{Should \code{\link{datatyper}} be run after reading?}

\item{drop_dim}{Should \code{\link{drop_img_dim}} be run after reading?}

\item{warn}{Should warnings from \code{\link{readNIfTI}} be 
printed?  If not, \code{\link{suppressWarnings}} is called.  Also passed to 
\code{\link{datatyper}}}
}
\value{
\code{nifti} object
}
\description{
This function calls the \code{\link{readNIfTI}} function from the 
\code{oro.nifti} package, but sets the reorientation to \code{FALSE} by default
}

