% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readrpi.R
\name{readrpi}
\alias{readrpi}
\title{Read NIfTI file reoriented to RPI}
\usage{
readrpi(file, ..., verbose = TRUE)
}
\arguments{
\item{file}{file name of the NIfTI file.}

\item{...}{Arguments to pass to \code{\link{readnii}}}

\item{verbose}{print diagnostics, passed to \code{\link{rpi_orient_file}}}
}
\description{
This function calls the \code{\link{readnii}} function after
calling \code{\link{rpi_orient_file}} to force RPI orientation.
}
