% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_window.R
\name{robust_window}
\alias{robust_window}
\title{Window image based on quantiles of Image}
\usage{
robust_window(img, probs = c(0, 0.999), ...)
}
\arguments{
\item{img}{object of class nifti}

\item{probs}{quantiles to constrain the image these define the window sent to \code{\link{window_img}}}

\item{...}{additional arguments sent to \code{\link{window_img}}}
}
\value{
Object of class nifti with values outside quantiles replaced
by values depending on replace argument passed to \code{\link{window_img}}
}
\description{
Takes an image, finds the quantiles given by 
\code{probs}, then sets values outside these values to other 
values, as determined by \code{replace} argument passed to
to \code{\link{window_img}}
}

