% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_orient.R
\name{rpi_orient}
\alias{rpi_orient}
\title{Reorient an Image to RPI orientation}
\usage{
rpi_orient(file, verbose = TRUE)
}
\arguments{
\item{file}{Object of class \code{nifti} or character path}

\item{verbose}{print diagnostic messages}
}
\value{
List of 3 elements
\itemize{
\item{\code{img}: }{Reoriented image of class \code{nifti}}
\item{\code{convention}: }{Convention (Neurological/Radiological) of original image}
\item{\code{orientation}: }{Original image orientations}
}
}
\description{
This function uses \code{fslswapdim} to reorient an image
}

