% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scl_inter.R
\docType{methods}
\name{scl_inter-methods}
\alias{scl_inter,character-method}
\alias{scl_inter-methods}
\title{Extract Image scl_inter attribute}
\usage{
\S4method{scl_inter}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
scl_inter method for character types
}

