% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scl_slope.R
\docType{methods}
\name{scl_slope}
\alias{scl_slope}
\title{Extract NIfTI 3D Image slope attribute}
\usage{
scl_slope(object)

\S4method{scl_slope}{nifti}(object)

scl_slope(object) <- value

\S4method{scl_slope}{nifti,numeric}(object) <- value

\S4method{scl_slope}{character}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to scl_slope}
}
\description{
Methods that act on the ``scl_slope'' numeric in the NIfTI header.
}

