% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scl_slope.R
\docType{methods}
\name{scl_slope-methods}
\alias{scl_slope-methods}
\alias{scl_slope,character-method}
\title{Extract Image scl_slope attribute}
\usage{
\S4method{scl_slope}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
scl_slope method for character types
}
