% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sform_code.R
\docType{methods}
\name{sform_code-methods}
\alias{sform_code,character-method}
\alias{sform_code-methods}
\title{Extract Image sform_code attribute}
\usage{
\S4method{sform_code}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
sform_code method for character types
}

