% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slope.R
\docType{methods}
\name{slope}
\alias{slope}
\title{Extract NIfTI 3D Image slope attribute}
\usage{
slope(object)

\S4method{slope}{nifti}(object)

slope(object) <- value

\S4method{slope}{nifti}(object) <- value

\S4method{slope}{character}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to slope}
}
\description{
Methods that act on the ``slope'' in the NIfTI header.
}

