% Generated by roxygen2 (4.0.2): do not edit by hand
\name{susan.help}
\alias{susan.help}
\title{FSL SUSAN Help}
\usage{
susan.help()
}
\value{
Prints help output and returns output as character vector
}
\description{
This function calls \code{susan}'s help
}
\examples{
if (have.fsl()){
 susan.help()
}
}

