% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tempimg}
\alias{tempimg}
\title{Create temporary nii.gz file for FSL}
\usage{
tempimg(nim)
}
\arguments{
\item{nim}{object of class nifti}
}
\value{
filename of output nii.gz
}
\description{
Takes in a object of class nifit, writes it to a temp file, appends
.nii.gz as \code{\link{writeNIfTI}} adds it.
}

