% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempimg.R
\name{tempimg}
\alias{tempimg}
\title{Create temporary nii.gz file for FSL}
\usage{
tempimg(nim, gzipped = TRUE, checknan = TRUE, check_type = FALSE,
  warn = FALSE, ...)
}
\arguments{
\item{nim}{object of class nifti}

\item{gzipped}{Should file be gzipped? Passed to 
\code{\link{writenii}}}

\item{checknan}{Check for NAs or NaNs}

\item{check_type}{Check the datatype for an image.  Will run \code{\link{datatyper}}.}

\item{warn}{Should warnings be displayed if \code{\link{writenii}} has
any?  Passed to \code{\link{writenii}}.}

\item{...}{Not used}
}
\value{
filename of output nii.gz
}
\description{
Takes in a object of class nifti, writes it 
to a temp file, appends
.nii.gz as \code{\link{writenii}} adds it.
}

