% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffset.R
\docType{methods}
\name{toffset-methods}
\alias{toffset-methods}
\alias{toffset,character-method}
\title{Extract Image toffset attribute}
\usage{
\S4method{toffset}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
Gets toffset from a character
}
