% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vox_offset.R
\docType{methods}
\name{vox_offset-methods}
\alias{vox_offset,character-method}
\alias{vox_offset-methods}
\title{Extract Image vox_offset attribute}
\usage{
\S4method{vox_offset}{character}(object)
}
\arguments{
\item{object}{is a filename to pass to \link{fslval}}
}
\description{
Extract Image vox_offset attribute
}

